Major Updates - more Screens Implemented for UI_V2 and finally after so many iterations Cart Activity is working as Intended, except for extremely Huge Load Testing. Screens Implemented :- 1 - LastFiveTxns Activity, txn_card.xml 2 - TxnDetails Activity, txn_details_card.xml 3 - AboutUs Activity 4 - ContactUS Activity 5 - Covid19 Activity 6 - IntroActivity 7 - MyProfile Activity 8 - NoQStores Activity 9 - PrivacyPolicy Activity 10 - ProductDetails Activity(Updated) 11 - Products Category Activity(Updated) 12 - ProductsList(Updated) 13 - RefundPolicy Activity 14 - SlideFirstFragment Activity 15 - SlideSecondFragment Activity 16 - SlideThirdFragment Activity 17 - Terms and Conditions Activity
1 - AwsBackgroundWorker
1_1 - added update_interested api, to update the no. of Users in either takeaway or home_delivery or both.
2 - BarcodeScannerActivity
2_1 - removed continuously popping Toast msg for Permission to access camera by NoQ.
2_2 - added shoppingMethod as putExtra() in Go_to_Basket() Func.
3 - CartActivity
3_1 - added coming_from, to retrieve from the Intent.
3_2 - removed total_items variable as its function was similar to item_qty.
3_3 - removeItem() Func.
3_3_1 - rearranged the Func. to be more organized.
3_3_2 - now the total_our_price is multiplied by qty, then subtracted from itself.(this is done as now removeItem removes the specified product with all qty's present, instead of removing just 1 qty).
3_4 - onCreate() Func.
3_4_1 - added checks to retrieve selective inputs from Intent, based on shoppingMethod.
3_4_2 - added onIncreaseQuantity() and onDecreaseQuantity() Funcs. to adapter.
3_5 - added onIncreaseQuantity() Func. to Increase the Quantity of the selected Product by 1.
3_6 - added onDecreaseQuantity()Func. to Decrease the Quantity of the selected Product by 1.
3_7 - added value for Quantity that is retrieved from the Database, to the Product Class.
3_8 - added Extra Checks in onBackPressed() and Go_Back() Funcs. in Takeaway ShoppingMethod Block, that if Intent is comingFrom 'ProductsCategory' then on pressing back it should be sent to ProductsCategory Screen else to ProductsList Screen.
3_9 - updated the value of item_qty being sent to Payment Success/FailedActivity in afterPaymentConfirm() Func.
3_10 - added Checks in onBackPressedCancelTransaction() based on shoppingMethod, that Cart was on before going to Paytm SDK.
4 - ChooseShopType Activity
4_1 - added code to Update the Interested Column in Store_Table in both Takeaway and Home_Delivery Shopping Methods.
4_2 - override onBackPressed() Func. to route the User back to MyProfile Screen.
5 - DBHelper.java
5_1 - added Quantity Field to Products_Table, and Updated both insertData() and insertProductData() Funcs.
5_2 - Updated update_product() Func. now it accepts quantity of the Product that has to be Updated in the Database.
5_3 - added update_quantity() Func. to increase or decrease the qty of the specified product, based on operation parameter.
5_4 - added getProductQuantity() Func. to retrieve the Product's Qty Present in the Database.
6 - LastFiveTxns Activity
6_1 - added referral_used parameter to the Txn Class.
7 - NoqStores Activity
7_1 - changed the names of the variables to something meaningful and added code to extract the Initials of the User's name and put in profile icon image.
8 - Payment Success/Failed Activity
8_1 - added Check for the Payment Method, so that if it is [Referral_Used], change it to Bonus.
9 - PaymentSuccess Activity
9_1 - added isDirectLogin flag in Go-to_Profile() Func. to ensure that the Bonus Banner doesn't pop up.
10 - Product.java
10_1 - added setter method for Current_qty.
11 - ProductAdapter.java
11_1 - added DBHelper for performing db related operations.
11_2 - added OnIncreaseQuantity() and onDecreaseQuantity() Funcs.
11_3 - removed resize(), centerCrop() and added fit(), so that Image can properly fit the ImageView.
11_4 - updated the names of the variables to something meaningful.
11_5 - added onClickListeners on im_add and im_delete to add and delete the Quantity of any Product, with robust checks.The changes are instantly updated in the Database.
11_6 - and now as soon as user clicks on add or delete button the Current Product qty in updated in ProductList also, tp avoid any recyclerview related Issues.
11_7 - Proper msg are given to the User.
12 - ProductDetails Activity
12_1 - Updated the Layout of the ProductList Items, now the Product status(In Stock/Out of Stock) and Current_Qty of the product(if present in the Database) is shown to the User.
12_2 - added comingFrom String, to make the Flow going smooth.
12_3 - added functionality to increase or decrease the Qty. of the Selected Item, and now the Item_qty is by default set to 1.
12_4 - added comingFrom String to add_more Func.
12_5 - added check so that if the Shopping Method is In Store then set the Add more Btn text to Scan More.
12_6 - updated onBackPressed() Func. to be more robust, based on shoppingMethod.
12_7 - now currently entered qty is also sent, to the update_product() Func.
12_8 - added checks, so that add to cart or increase/decrease option are visible to users only when the available qty of the product is > 0.If not show Out_of_stock.
12_9 -  and after successfully adding/updating product to the local database, the prod_qty count is set to 1.
13 - ProductListAdapter.java
13_1 - now each list item has an option to increase or decrease the qty of the Product.
13_2 - now if the product already exists in the local database, then the Current Qty is retrieved from the database.
13_3 - and after successfully adding/updating product to the local database, the prod_qty count is set to 1.
13_4 - added checks to show or hide the options to add product to cart or increase or decrease qty of the product, if available qty greater/less than 0 resp.If < 0 show Out_of_stock.
13_5 - now product's category and Quantity is also added to the List being sent to the Database for Storing/Updating.
13_6 - added Quantity field to the list being sent to ProductDetails Screen.
14 - ProductsCategory Activity
14_1 - added Cart option for uses.
15 - ProductsList Activity
15_1 - added comingFrom String to Go_to_Basket() Func.
15_2 - now user is sent directly back to ProductsCategory Screen when clicked onBackPressed Fun().
16 - Txn.java
16_1 - added referral_used parameter to the Class.
17 - TxnAdapter.java
17_1 - made some UI changes, and added/updated the code for that in this.
17_2 - now the time shown is in 12hr format.
18 - TxnDetails Activity
18_1 - now the time shown is in 12hr format.
18_2 - added you_saved field.
18_3 - added Check for the Payment Method, so that if it is [Referral_Used], change it to Bonus.
19 - TxnDetailsAdapter.java
19_1 - Updated the UI for this Screen.
20 - added some more icons/illustrations for the UI_V2.

+++ NoQ_1/app/src/main/java/com/younoq/noq/CartActivity.java
@@ -49,7 +49,7 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
     SaveInfoLocally save;
     Bundle txnReceipt;
     ArrayList<String> txnData;
-    private String shoppingMethod, category_name;
+    private String shoppingMethod, category_name, coming_from;
 
     JSONArray jsonArray;
     JSONObject jobj1, jobj2;
@@ -68,41 +68,31 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
     private double total_our_price = 0.0;
     private double total_discount = 0.0;
     private static int item_qty = 0;
-    private int total_items = 0;
     private DecimalFormat df = new DecimalFormat("###.##");
     String txnAmount;
     private String comment = "";
     private static final String TAG = "CartActivity";
 
     // Function to Delete a Specific Item from the Basket, based on its position in the cart.
-    public void removeItem(int position, int id, double our_price, double mrp, int qty, double retail_price, double tot_discount){
-        // Removing the Product from the ProductList.
-        ProductList.remove(position);
-        // Notifying the Adapter of the Change.
-        adapter.notifyItemRemoved(position);
-        // Creating an Instance of the DB.
+    public void removeItem(int position, int id, double our_price, double mrp, int qty, double retail_price, double tot_discount, boolean delete){
+
         dbHelper = new DBHelper(this);
 
         // Reducing the MRP(total_mrp * qty) of the Item that is being deleted, from Total_MRP.
-//        Log.d(TAG, "Total_MRP Before Reduction : "+total_mrp+" Item_MRP : "+mrp+" Item Qty : "+qty);
         total_mrp -= (mrp * qty);
-//        Log.d(TAG, "Total_MRP After Reduction : "+total_mrp);
 
         //Reducing the Total No. of Items in the Cart
         item_qty -= qty;
-        items_qty.setText(String.valueOf(item_qty));
+        final String tot_items = "("+item_qty+")";
+        items_qty.setText(tot_items);
 
         // Reducing the Total_Retailers_Price (retail_price * qty) of the Item that is being deleted, from Total_Retail_Price.
-//        Log.d(TAG, "Total_Retail_Price Before Reduction : "+total_retail_price+" Item_MRP : "+retail_price+" Item Qty : "+qty);
         total_retail_price -= (retail_price * qty);
-//        Log.d(TAG, "Total_Retail_Price After Reduction : "+total_retail_price);
 
         // Reducing the Total_Discount (tot_discount * qty) of the Item that is being deleted, from Total_Discount.
-//        Log.d(TAG, "Total_Discount Before Reduction : "+total_discount+" Item_MRP : "+tot_discount+" Item Qty : "+qty);
         total_discount -= (tot_discount * qty);
-//        Log.d(TAG, "Total_Discount After Reduction : "+total_discount);
+
         total_discount = Double.parseDouble(df.format(total_discount));
-//        Log.d(TAG, "Total_Discount After Reduction using DecimalFormat : "+total_discount);
 
         // --------------------------------------- If Referral Enabled -------------------------------
         // Current Final_Amount Value.
@@ -111,11 +101,10 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
         // --------------------------------------- X X X X X X X X X X X -----------------------------
         if(total_our_price > 0.0){
 
-//            Log.d(TAG, "Total_Our_Price Before Reduction : "+total_our_price+" Item_Price : "+our_price);
-            total_our_price -=  our_price;
-//            Log.d(TAG, "Total_Our_Price After Reduction : "+total_our_price);
+            total_our_price -=  (our_price * qty);
+
             total_our_price = Double.valueOf(df.format(total_our_price));
-//            Log.d(TAG, "Total_Our_Price After Reduction using DecimalFormat : "+total_our_price);
+
 
             final Double value_ref_bal =  Double.valueOf(ref_bal);
             if ( total_our_price > value_ref_bal ) {
@@ -124,7 +113,6 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
                 current_final_amt = "0";
             }
         } else {
-//            Toast.makeText(this, "Kindly Revisit the Page..", Toast.LENGTH_SHORT).show();
             current_final_amt = "0";
         }
         if(total_our_price == 0.0){
@@ -139,7 +127,15 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
         tv_final_amt.setText(amt1);
         // -------------------------------- X X X X X X X X X X X -------------------------------------
         // Deleting the Specific Product from the DB.
-        dbHelper.DeleteData_by_id(id);
+        Log.d(TAG, "Delete Product : "+delete);
+        if(delete){
+            dbHelper.DeleteData_by_id(id);
+            // Removing the Product from the ProductList.
+            ProductList.remove(position);
+            // Notifying the Adapter of the Change.
+            adapter.notifyItemRemoved(position);
+            // Creating an Instance of the DB.
+        }
     }
 
     @Override
@@ -169,7 +165,9 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
         Intent in = getIntent();
         shoppingMethod = in.getStringExtra("shoppingMethod");
         if(shoppingMethod.equals("Takeaway")){
+            coming_from = in.getStringExtra("comingFrom");
             category_name = in.getStringExtra("category_name");
+            scan_product.setText(R.string.ca_add_more);
         }
 
         RetrieveFromDatabase();
@@ -179,17 +177,22 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
         // If any product is Deleted from the Cart, to retrieve the details of the deleted item.
         adapter.setOnItemClickListener(new ProductAdapter.onItemClickListener() {
             @Override
-            public void onDeleteClick(int position, int id, double our_price, double mrp, int qty, double retail_price, double tot_discount) {
+            public void onDeleteClick(int position, int id, double our_price, double mrp, int qty, double retail_price, double tot_discount, boolean delete) {
                 Toast.makeText(CartActivity.this, "Item Deleted", Toast.LENGTH_SHORT).show();
                 Log.d(TAG, "Position : "+position+" Id : "+id+" Price : "+our_price);
-                removeItem(position, id, our_price, mrp, qty, retail_price, tot_discount);
+                removeItem(position, id, our_price, mrp, qty, retail_price, tot_discount, delete);
 //                RetrieveFromDatabase();
             }
 
-//            @Override
-//            public void onItemClick(int position) {
-//
-//            }
+            @Override
+            public void onIncreaseQuantity(int position, int id, double mrp, double our_price, double retail_price, double discount) {
+                increaseQuantity(position, id, mrp, our_price, retail_price, discount);
+            }
+
+            @Override
+            public void onDecreaseQuantity(int position, int id, double mrp, double our_price, double retail_price, double discount, boolean delete) {
+                decreaseQuantity(position, id, mrp, our_price, retail_price, discount, delete);
+            }
         });
 
         // if User clicks on Scan_Product Button in UI.
@@ -216,6 +219,132 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
         // Fetch the User's Referral Balance
         fetch_referral_amt();
         // -------------------------------- X X X X X X X X X X X ---------------------------------------
+
+    }
+
+    private void increaseQuantity(int position, int id, double mrp, double our_price, double retail_price, double discount) {
+
+        // Initializing DBHelper
+        dbHelper = new DBHelper(this);
+
+        final String details = "Before -> pos : "+position+", id : "+id+", mrp : "+mrp+", our_price : "+our_price+", retail_price : "+retail_price;
+        final String amt2 = "Total_our_Price : "+total_our_price+", Total_MRP : "+total_mrp;
+        Log.d(TAG, details);
+        Log.d(TAG, amt2);
+
+        final int qty = 1;
+
+        // Increasing the MRP(total_mrp * qty) for the Selected Item.
+        total_mrp += (mrp * qty);
+
+        //Increasing the Total No. of Items in the Cart
+        item_qty += qty;
+        final String tot_items = "("+item_qty+")";
+        items_qty.setText(tot_items);
+
+        // Increasing the Total_Retailers_Price (retail_price * qty) of the Item.
+        total_retail_price += (retail_price * qty);
+
+        // Increasing the Total_Discount (tot_discount * qty) of the Item.
+        total_discount += (discount * qty);
+
+        total_discount = Double.parseDouble(df.format(total_discount));
+
+        // --------------------------------------- If Referral Enabled -------------------------------
+        // Current Final_Amount Value.
+        String current_final_amt = tv_final_amt.getText().toString();
+        current_final_amt = current_final_amt.replace("₹", "");
+        // --------------------------------------- X X X X X X X X X X X -----------------------------
+        total_our_price +=  our_price;
+
+        total_our_price = Double.valueOf(df.format(total_our_price));
+
+        final Double value_ref_bal =  Double.valueOf(ref_bal);
+
+        if ( total_our_price > value_ref_bal ) {
+            current_final_amt = String.valueOf(total_our_price - value_ref_bal);
+        } else {
+            current_final_amt = "0";
+        }
+
+        final String amt = "₹"+df.format(total_our_price);
+        tv_total_our_price.setText(amt);
+        // ------------------------------- If Referral Enabled ---------------------------------------
+        final String amt1 = "₹"+ df.format(Double.valueOf(current_final_amt));
+        tv_final_amt.setText(amt1);
+        // -------------------------------- X X X X X X X X X X X -------------------------------------
+        final String details1 = "After -> pos : "+position+", id : "+id+", mrp : "+mrp+", our_price : "+our_price+", retail_price : "+retail_price;
+        final String amt3 = "Total_our_Price : "+total_our_price+", Total_MRP : "+total_mrp;
+        Log.d(TAG, details1);
+        Log.d(TAG, amt3);
+
+    }
+
+    private void decreaseQuantity(int position, int id, double mrp, double our_price, double retail_price, double discount, boolean delete) {
+
+        // Initializing DBHelper
+        dbHelper = new DBHelper(this);
+
+        final int qty = 1;
+
+        // Reducing/Adding the MRP(total_mrp * qty) for the Selected Item, from Total_MRP.
+        total_mrp -= (mrp * qty);
+
+        //Reducing the Total No. of Items in the Cart
+        item_qty -= qty;
+        final String tot_items = "("+item_qty+")";
+        items_qty.setText(tot_items);
+
+        // Reducing the Total_Retailers_Price (retail_price * qty) of the Item that is being deleted, from Total_Retail_Price.
+        total_retail_price -= (retail_price * qty);
+
+        // Reducing the Total_Discount (tot_discount * qty) of the Item that is being deleted, from Total_Discount.
+        total_discount -= (discount * qty);
+
+        total_discount = Double.parseDouble(df.format(total_discount));
+
+        // --------------------------------------- If Referral Enabled -------------------------------
+        // Current Final_Amount Value.
+        String current_final_amt = tv_final_amt.getText().toString();
+        current_final_amt = current_final_amt.replace("₹", "");
+        // --------------------------------------- X X X X X X X X X X X -----------------------------
+        if(total_our_price > 0.0){
+
+            total_our_price -=  our_price;
+
+            total_our_price = Double.valueOf(df.format(total_our_price));
+
+            final Double value_ref_bal =  Double.valueOf(ref_bal);
+            if ( total_our_price > value_ref_bal ) {
+                current_final_amt = String.valueOf(total_our_price - value_ref_bal);
+            } else {
+                current_final_amt = "0";
+            }
+        } else {
+            current_final_amt = "0";
+        }
+        if(total_our_price == 0.0){
+            payment_btn.setVisibility(View.INVISIBLE);
+        } else {
+            payment_btn.setVisibility(View.VISIBLE);
+        }
+        final String amt = "₹"+df.format(total_our_price);
+        tv_total_our_price.setText(amt);
+        // ------------------------------- If Referral Enabled ---------------------------------------
+        final String amt1 = "₹"+ df.format(Double.valueOf(current_final_amt));
+        tv_final_amt.setText(amt1);
+        // -------------------------------- X X X X X X X X X X X -------------------------------------
+        // Deleting the Specific Product from the DB.
+        Log.d(TAG, "Delete Product : "+delete);
+        if(delete){
+            dbHelper.DeleteData_by_id(id);
+            // Removing the Product from the ProductList.
+            ProductList.remove(position);
+            // Notifying the Adapter of the Change.
+            adapter.notifyItemRemoved(position);
+            // Creating an Instance of the DB.
+            dbHelper.close();
+        }
 
     }
 
@@ -272,8 +401,6 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
                     final double qty = Double.parseDouble(res.getString(3));
                     // Setting the Total No. of Items in the Database
                     item_qty += qty;
-                    // Calculating Total Items in the Cart.
-                    total_items += qty;
     //                Log.d(TAG, "Quantity : "+qty);
                     // Retrieving the (No.of Items * Total_MRP) from the Database for all the Entries.
                     total_mrp += qty * Double.parseDouble(res.getString(5));
@@ -299,7 +426,7 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
                                     res.getString(9),
                                     res.getString(3),
                                     res.getString(10),
-                                    "0",
+                                    res.getString(11),
                                     ""
                             )
                     );
@@ -319,7 +446,8 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
 
         final String amt = "₹"+df.format(total_our_price);
         tv_total_our_price.setText(amt);
-        items_qty.setText(String.valueOf(item_qty));
+        final String tot_items = "("+item_qty+")";
+        items_qty.setText(tot_items);
 
         adapter = new ProductAdapter(this, ProductList);
         recyclerView.setAdapter(adapter);
@@ -355,6 +483,7 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
     protected void onStop() {
         super.onStop();
         total_our_price = 0.0;
+        item_qty = 0;
     }
 
     @Override
@@ -365,10 +494,15 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
             in.putExtra("Type", "Product_Scan");
             in.putExtra("shoppingMethod", shoppingMethod);
         } else if(shoppingMethod.equals("Takeaway")){
-            in  = new Intent(CartActivity.this, ProductsList.class);
-            in.putExtra("coming_from", "Cart");
+            // If coming from Products Category Screen then, go back there.
+            if(coming_from.equals("ProductCategory")){
+                in  = new Intent(CartActivity.this, ProductsCategory.class);
+            } else {
+                in  = new Intent(CartActivity.this, ProductsList.class);
+                in.putExtra("coming_from", "Cart");
+                in.putExtra("category_name", category_name);
+            }
             in.putExtra("shoppingMethod", shoppingMethod);
-            in.putExtra("category_name", category_name);
         } else{
             in = new Intent();
         }
@@ -382,15 +516,18 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
         if(shoppingMethod.equals("In Store")){
             in  = new Intent(CartActivity.this, BarcodeScannerActivity.class);
             in.putExtra("Type", "Product_Scan");
-            in.putExtra("shoppingMethod", shoppingMethod);
         } else if(shoppingMethod.equals("Takeaway")){
-            in  = new Intent(CartActivity.this, ProductsList.class);
-            in.putExtra("coming_from", "Cart");
-            in.putExtra("shoppingMethod", shoppingMethod);
-            in.putExtra("category_name", category_name);
+            if(coming_from.equals("ProductCategory")){
+                in  = new Intent(CartActivity.this, ProductsCategory.class);
+            } else {
+                in  = new Intent(CartActivity.this, ProductsList.class);
+                in.putExtra("coming_from", "Cart");
+                in.putExtra("category_name", category_name);
+            }
         } else{
             in = new Intent();
         }
+        in.putExtra("shoppingMethod", shoppingMethod);
         in.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
         startActivity(in);
 
@@ -513,7 +650,7 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
                     txnData.add(final_user_amt);
                     txnData.add(time);
                     txnData.add(Pay_Mode);
-                    txnData.add(String.valueOf(total_items));
+                    txnData.add(String.valueOf(item_qty));
                     // Adding the txnData ArrayList to txnReceipt Bundle.
                     txnReceipt.putStringArrayList("txnReceipt", txnData);
                     Log.d(TAG, "Stored Required Details in Bundle");
@@ -721,9 +858,26 @@ public class CartActivity extends AppCompatActivity implements PaytmPaymentTrans
 
     @Override
     public void onBackPressedCancelTransaction() {
-//        Toast.makeText(this, "Back Pressed", Toast.LENGTH_LONG).show();
-        Intent in = new Intent(this, CartActivity.class);
+//        Toast.makeText(this, "Back Pressed", Toast.LENGTH_LONG).show();Intent in;
+        Intent in;
+        if(shoppingMethod.equals("In Store")){
+            in  = new Intent(this, CartActivity.class);
+            in.putExtra("shoppingMethod", shoppingMethod);
+        } else if(shoppingMethod.equals("Takeaway")){
+            in  = new Intent(this, CartActivity.class);
+            in.putExtra("comingFrom", "Cart");
+            in.putExtra("shoppingMethod", shoppingMethod);
+            in.putExtra("category_name", category_name);
+        } else{
+            in = new Intent();
+        }
+//        in.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
         startActivity(in);
+
+//        Intent in = new Intent(this, CartActivity.class);
+//        in.putExtra("shoppingMethod", shoppingMethod);
+//        in.putExtra("category_name", category_name);
+//        startActivity(in);
     }
 
     @Override
